epic performance boost
remove the null byte check; originally the code was scanning each line
for a NUL with regexp.Match() - compiling the regexp in the function.
This compilation turned out to be rather expensive, not to mention the
added cost for matching regexps on each line. The updated behaviour of
only looking for specific files largely obviates the need for this.